[simple-app] Make the chart "connection" test configurable (#8)
**What did I want?**

The existing test-case was hard coded to a plain-text HTTP get with `wget` as the client. That doesn't make it easy to support alternative use-cases (gRPC for example).

**What did I do?**

Reconfigured the test suite so that all of the details are in the `values.yaml` file under the `tests.connection` key. This not only makes the test more configurable, but it also lets us start adding in other tests down the road that might be different and gives us independent control to configure them.