[istio-endpoints] Create a chart for managing per-namespace Istio configurations (#14)
**What did I want?**

Managing Istio configurations is complicated and risky. I want to centralize the management of the configurations for different service-types so that we can have some consistency.

**What did I do?**

I have created a new `istio-endpoints` chart that manages the entire Istio sidecar configuration for a given Namespace. The motivation here is to not only allow developers to use this chart directly, but to also use this chart in the `simple-app` as an optional dependency chart.