[prometheus-alerts] Fix the OOMKilled alert so that it sum's properly (#38)
**What did I want?**

I discovered that we had pods being OOMed and we were not being alerted. It turns out that the original `PodContainerTerminated` alert was looking for a container that was permanently in the OOMKilled/Error/ContainerCannotRun state.. but since containers get restarted within a Pod, we need to
monitor for the _count_ of that metric, not the _current state_ of it.

Here is the example of the "view" that the original alert had on a pod that was OOMing every few mins:
![image](https://user-images.githubusercontent.com/768067/129816039-c9fce8fc-20b7-4ba4-86a2-3b2259b4543d.png)

Obviously this was not going to trip the alarm.

**What did I do?**

I replaced the query with one that sum's over time. The `for` value is now used as the "sum over this period" value, and the alarm trips as soon as it trips. Here is the graph view of this new alarm:

![image](https://user-images.githubusercontent.com/768067/129816093-75a489d3-6175-495b-88df-cedbe71c4c07.png)

While I was here, I limit the number of labels returned so that the alarm is more clear.