[simple/daemon/stateful-apps] Support overriding the image Tag manually (#89)
When using ArgoCD and the ArgoCD Image Updater, we want to allow the
image updater to do its job (setting `.image.tag`), but also allow for
humans to override the updater. Creating a new `image.forceTag` setting
lets a human with access to the `Application`. The argocd-image-updater
will not have any idea that it's change is being overridden, and this
new setting makes it more clear to application owners that a
human-introduced change has been applied.