[simple-app] Import all of the relevant PrometheusRules from the prometheus-alerts chart
The goal of this PR is to make sure that each time a `simple-app` chart
is launched, it automatically includes relevant alerts for the
application itself. Many of these alerts are tricky because we have to
use `pod=~...` regex matches to scope things correctly.

The intention is that `simple-app` here will only have alerts relevant
to its own configuration... so if a chart uses several `simple-app`
charts, we'll end up with a different set of Prometheus Rules for each
chart.