[istio-alerts] Introduce .Chart.Name into Rule Names
The motivation of introducing another field into .Chart.Name here is to
make it easier to identify where the "ServiceRules" are coming from -
without this, they appear to be natively part of the parent-chart
(simple-app for example). With the .Chart.Name in there, you can more
easily see that these come from a nested chart.