[daemonset-app] Add in a verticalAutoscaling configuration (#107)
This PR adds in management of a `VerticalPodAutoscaler` resource for the `daemonset-app` chart. I intend to add this to some of the other charts down the road, but the `daemonset-app` chart makes the most sense right now because we have a few daemonsets where the resources are very much situationally dependent ... and trying to tune the resource requests by hand is unreasonable.