[simple-app/stateful-app] Increase the Minimum Replica Count and set a PDB (#117)
Our devlopers are (rightfully) not setting `replicaCount` or
`podDisruptionBudget` configurations in staging because they haven't
felt the need to. However this means that during node replacements and
even during code rollouts we get the ArgoCD Notifications App reporting
that the service is "failed" just because something like the HPA wasn't
able to find metrics for a pod during the rollout.

It is reasonable to make our minimum default deployment include at least
2 pods, and have a `PodDisruptionBudget` created that allows only one
pod to be unavailable at a time. This more accurately reflects the
behaviors we have in production anyways.