[apps] Refactor the Datadog Labels/Env variables into nd-common
These are common across all of our apps... so I am refactoring them into
the nd-common library chart so they are easier to re-use. Additionally,
we are making the DD_ENV variable optional and defaulting it to `null`
so that it is not set.

If you set the DD_ENV variable, it overrides any "env" value that is
configured in the underlying host datadog agent configuration. We
actually want our developers to not think about this at all, and let the
actual computing environment determine the "env" tag assigned to their
traces and metrics.