[rollout-app] New Argo Rollouts based Deployment (#136)
**What did I want?**

As we begin using [Argo Rollouts](https://argoproj.github.io/argo-rollouts/) - we need a new default application template that can create the `Rollout` objects, as well as the more customized `VirtualService` and `Service` resources. I wanted to enable us to begin experimenting with these deployment models in a simple way.

**What did I do?**

A newly created `rollout-app` chart has been created along with some customization to the Github Workflow. The workflow now detects if the `rollout-app` chart was updated, and if it was, we then install the Argo Rollouts controller/CRD for the tests. When the tests run, we verify that we can create both a `BlueGreen` and `Canary` rollout resource type. We are explicitly not testing actually triggering a rollout - we don't need to test the fundamental behavior of Argo, just that our chart basically works.