[stateful/simple/nd-common] Implement a smarter TopologySpreadConstraint configuration
The original setup allowed you to pass in a `TopologySpreadConstraints`
list of maps ... but we did no template parsing, no validation, and we
provided no intelligent way to set the `labelSelector` key.

This PR does two things. First we intelligently parse the
`TopologySpreadConstraints` maps that the user supplies, error out
appropriately, and then insert the `labelSelector` automatically.

Second, we create a default topology spread constraint that spreads
across AZs, allowing users to just flip a boolean flag and get the
spread done in the simplest possible way.