[simple-app/stateful-app/daemonset-app] Use prometheus-alerts chart (#154)
I realized that at one point we started going down the path of breaking
out the different `PrometheusRules` into different charts (eg, moving
"Deployment" rules into the `simple-app` chart). The problem is that it
made it hard to reuse these rules elsewhere ... and we started to lose
sync of them between the `prometheus-rules` chart and the various
application charts.

Instead, this PR goes back to automatically including the
`prometheus-alerts` chart just like the `istio-alerts` chart is
included. The appropriate rules are disabled in each chart so that we
only create the rules that are necessary.