[simple/daemon/stateful] Revert to using in-template prometheusrules (#159)
* [simple-app/stateful-app] Remove KubeDaemonSetRolloutStuck alarm

* [simple/daemon/statefulset] Revert to using in-template prometheusrules

The nested-nested-nested chart system in Helm is really quite bad and it
is really impossible to build nested charts with tons of nested charts
under them. You run into all kinds of naming conflict issues sadly.

Instead, I am going the other way and finishing migrating all of the
PrometheusRules into each of the DaemonSet-App, Stateful-App and
Simple-App charts. When this is done, we will sadly need to maintain
both the prometheus-alerts chart, AND these other charts .. but it's
just not easy to deal with this any other way right now.