[simple-app/statefull-app/daemonset-app] Always use a Service (#160)
**What did I want?**

We discovered
(https://github.com/istio/istio/issues/39792#issuecomment-1189669761)
that Istio requires a `Service` point to all workloads (clients and
servers) in order to generate their "locality" information. Without a
`Service` pointing to the workload, Istio will not  generate this
locality information and we cannot leverage zone-aware routing.

**What did I do?**

All of our apps now create a `Service` resource no matter what that
points to the pods that are running. The `Service` does not need to be
exposed (meaning that no `NetworkPolicy` is created if the
`.Values.ports` key is empty), but we create the `Service` anyways so
that Istio can determine locality for all workloads.