feat(istio-alerts): Add service selector validity alert (#287)
> [!WARNING]
> Reminder for @LaikaN57 - Update other charts in this repo after this
PR is merged.

Split from https://github.com/Nextdoor/k8s-charts/pull/285

We ran into an issue where unbeknownst to us, our pod selector was set
incorrectly and therefore all of our rules deployed via
prometheus-alerts did not have any series to evaluate. This lack of
series data is silently ignored. Here we are resolving this by
implementing a rule which will use the various selectors to see if a
basic metric exists or not. If it does not exist then we want to warn
the user that non of the other rules will be evaluated.

Snippet from the other PR about using `kube_service_info` as our test
metric.

> I think I see what you're doing here ... but I see a risk. If there is
no traffic to the service for example, when its being launched), there
will be no metrics that match this query. I propose that instead we do
this:
>
> ```
> count(kube_service_info{namespace={{ $.Release.Namespace }}, name={{
$.Release.Namespace $.Values.serviceRules }})
> ```
>
> Sadly - I know that this means we aren't using the exact selectors
(like I propose above) ... but this allows us to verify that the Service
actually exists by that name. Ya know?

Lastly, I set the timing of this metric up for 1 hour by default. This
felt okay to me as we want to be a little bit resilient to prometheus
metric collection outages and not accidentally page every service owner
if we have a centralized outage of the platform itself. Normally, a team
might only be paged once for this when they first setup their
application if they had not setup their selectors correctly. This alert
is not expected to go into an alert state without some heavy handed
naming changes (mostly done during things like migrations).