feat(charts/istio-alerts): Remove source service from 5XX Alarm (#291)
As a follow up to https://github.com/Nextdoor/k8s-charts/pull/290 this
change actually adjusts the alarm to remove the source workload from the
grouping keys.

As a service owner, we want to provide out of the box visibility to the
health of the service with appropriate false positives and false
negatives. The current approach can create a high rate of false
positives in the case of a very low volume source service intermingled
with a higher volume source. This is currently happening with the
content service which has 99.9% of traffic from the monolith and a small
fraction of traffic via Bender, and was the impetus for the change.

```diff
➜  istio-alerts git:(fix-service-alarm-10) ✗ make helm template >> new.yaml
install.go:218: [debug] Original chart version: ""
install.go:235: [debug] CHART PATH: /Users/cory/src/k8s-charts/charts/istio-alerts

➜  istio-alerts git:(fix-service-alarm-10) ✗ diff -bu old.yaml new.yaml
--- old.yaml	2024-04-17 16:51:22
+++ new.yaml	2024-04-17 16:53:18
@@ -49,6 +49,7 @@
   groups:
   - name: observability.istio-alerts.istio-alerts.IstioServiceRules
     rules:
+
     - alert: 5xx-Rate-Too-High
       annotations:
         summary: >-
@@ -58,12 +59,12 @@
           High rate of 5xx responses from the {{$labels.destination_service_name}} VirtualService
           in namespace {{$labels.namespace}}.
       expr: >
-        sum by (destination_service_name, reporter, source_workload) (
+        sum by (destination_service_name, reporter) (
           istio_requests:increase5m{
             response_code=~"5.*",destination_service_name=~".*", destination_service_namespace="observability"}
         )
           /
-        sum by (destination_service_name, reporter, source_workload) (
+        sum by (destination_service_name, reporter) (
           istio_requests:increase5m{destination_service_name=~".*", destination_service_namespace="observability"}
         )
           > 0.0005
```