feat(prometheus-alerts): Add selector validity alarms (#297)
We ran into an issue where unbeknownst to us, our pod selector was set
incorrectly and therefore all of our rules deployed via
prometheus-alerts did not have any series to evaluate. This lack of
series data is silently ignored. Here we are resolving this by
implementing a rule which will use the various selectors to see if a
basic metric exists or not. If it does not exist then we want to warn
the user that none of the other rules will be evaluated.

I tried to choose metrics for each selector that were generic enough and
did not rely on a specific metric that may not be gathered. Most types
had an associated info or label metric which are good candidates for
this use-case.

Lastly, I set the timing of this metric up for 1 hour by default. This
felt okay to me as we want to be a little bit resilient to prometheus
metric collection outages and not accidentally page every service owner
if we have a centralized outage of the platform itself. Normally, a team
might only be paged once for this when they first setup their
application if they had not setup their selectors correctly. This alert
is not expected to go into an alert state without some heavy handed
naming changes (mostly done during things like migrations).

See https://github.com/Nextdoor/k8s-charts/pull/285