fix(prometheus-alerts): Make resource selectors more generic (#298)
We have changed the default selectors to be more generic. This means
that if you deploy multiple applications to the same namespace, and one
of those applications uses this chart, then by default all applications
will be monitored by these alerts. You can change this behavior by
modifying the `.Values.defaults.*NameSelector` regex values.

This is a more advantageous model as (like with the individual alerts
themselves) it now requires opt-out (or more specific scoping) if the
default is alerting on some objects you do not want to alert on.

Historical context:
- https://github.com/Nextdoor/k8s-charts/pull/149
- https://github.com/Nextdoor/k8s-charts/pull/150