feat(local-kafka): switch to kraft, start faster, allow annotations (#355)
## Goal

The goal for this PR is to improve the usability of the `local-kafka`
chart by reducing its startup time, introducing podAnnotations and
removing old legacy settings:

### Change: Removed legacy startup/readiness probe delays

We no longer see the Strimzi Operator take minutes to start up on M1
laptops, so I am removing the arbitrary startup/readiness probe delays.

### Change: Replace Zookeeper with KRaft

Strimzi is moving to removing Zookeeper entirely in the 0.46.0 release
anyways ... this change gets our local testing setup that way ahead of
the change. This also reduces the number of pods that have to run
locally, reducing the total startup time of a full testing environment.

### Change: Introduce `podAnnotations` and `annotations` settings

If you set `.Values.podAnnotations: { ... }` now we will add these
annotations to the Kafka and Entity Operator pods. This makes it easier
to tell Helm to start these pods up before the rest of the tested
resources get created.

---------

Co-authored-by: Matt Wise <diranged@users.noreply.github.com>