feat: integrated capabilities for multiple zones into charts/rollout-app (#384)
> [!IMPORTANT]
> For reviewers, I suggest going over the `rollout.yaml` file first to
get a better understanding of how zones are handled, what is checked,
etc. so that when reviewing the diffs to the other templates, things
make a bit more sense 🙂
## Context | Description | Why
Currently, if a user wanted to add multiple zones / add a Rollout
operator for different clusters, they would have to have multiple
instances of a Rollout CRD definition within a file / have multiple
files for each rollout on top of having to manually change and keep
track of certain fields within the `spec` in order to customize and make
each rollout service have different selectors.

This PR changes that, integrating the field `rolloutZones` to
automatically make Rollout operators for each specific zone, creating
duplicate rollout services for each one with other configurable options.
This is handy to have, for instance, if one of your configured zones
handles _substantially_ more traffic at times, and thus, can make it
such that splitting up the rollout between zones can help save some
costs.

## Testing
* running helm template showed no errors!
* I `helm install`'ed two versions of this rollout-app: one with an
empty `rolloutZones: []`, and it worked, and I also installed a version
of the app where the rolloutZones was `rolloutZones: [us-west-2a]`,
added an appropriate label to a node selector, and it worked! So,
rollout zones may need more configuration to get the node selection to
work, but otherwise, this should _not_ be a breaking change
* I also tested out the `rolloutZonesTransition` variable, and it
worked! Below is an example of how it works
* I also tested this out within the `dev1-test` cluster with a dummy
repo on a branch. To see the branch and testing, contact me so that I
can send it via slack! (since this is a public repo, I don't want to
share it here, haha)

## Videos
Contact me so that I can send the video of it working as well through
Slack!

---------

Signed-off-by: Eduardo Morales <emorales@nextdoor.com>