chore(charts/rollout-app): support dynamicStableScale option in canary (#386)
Adding support for `dynamicStableScale` in canary:
https://argo-rollouts.readthedocs.io/en/stable/features/canary/#dynamic-stable-scale-with-traffic-routing

When set, the rollout dynamically reduces the scale of the **stable**
ReplicaSet during an update such that it scales down as the traffic
weight increases to canary.

Advantage include:

1.  savings on spinup of new nodes in high-replica count environments
2. less likely for HPAs to scale stable down as overall pod counts stay
the same for similar traffic split when canary is being rolled out

<sub>By default the stable ReplicaSet is left scaled to 100% during the
update. This has the advantage that if an abort occurs, traffic can be
immediately shifted back to the stable ReplicaSet without delay (if set
to true, and the rollout is aborted, the canary ReplicaSet will
dynamically scale down as traffic shifts back to stable</sub>