fix(charts/simple-app): Disambiguate node selectors for transition deployments and per az (#391)
We currently see an error like "AmbiguousSelector pods by selector _ are
controlled by multiple HPAs" when setting per az transition to true.
This error is because the selector for the transition deployment is a
superset of the per az deployments. This pr adds an additional selector
to the transition deployment to make it no longer a superset of the per
az deployments