fix(charts/simple-app): fix bug in make template for transition (#395)
Make template prior to this pr had an issue where 
` error converting YAML to JSON: yaml: line 15: found character that
cannot start any token` when deployment zone transition was enabled. I
had tested make template with that value set to false but didn't test
that it worked when value was set to true. I have verified on this pr
that make template works and that the values for transition are correct
in the generated template