fix(nd-common): ensure consistent DD service names (#399)
> [!NOTE]
> This is a **chained** PR that should be merged in first before #398.
This sets up that PR to properly allow potential DataDog agents to
scrape any `k8s-charts.git` users Canary pods for logs.

## Context
For people using `k8s-charts.git` who use `rollout-app` and want to use
DataDog for observability, the service name for the DataDog annotations
is not aligned with what the _actual_ service names are. For example,
within `simple-app`, the name value for `Service` kinds is:
`nd-common.serviceName`. Thus, [when the DataDog Agent is trying to
scrape for
logs](https://docs.datadoghq.com/integrations/argo-rollouts/#log-collection),
the Agent will not be able to find the correct `<LOG_CONFIG>` values.
For `rollout-app`, the name is also changed to be aligned with
`simple-app` in the chained PR, but that is handled in the next PR that
follows this.

The solution for this is simply to ensure that the `<LOG_CONFIG>`
annotations for `ad.datadoghq.com/<CONTAINER_NAME>.logs` is aligned with
what is found everywhere else in the repo. This is more of a fix to
ensure future compatibility and minimize future confusion.

---------

Signed-off-by: Eduardo Morales <emorales@nextdoor.com>