fix(rollout-app): removed stable and canary pod level metadata annotations (#401)
## Context
The Argo Rollouts documentation states that pods can have tags with
values attached to them depending on their annotations, so I added
annotations to stable pods and canary pods that could be used to
differentiate between the two replicasets for DataDog dashboards.
However, I realized that the metrics that the DataDog agent can scrape
is also tied to the Argo Rollouts controller conf.yaml file, so if you
would want to use the metrics given by the Argo Rollouts x DataDog
integration, you need to specify it within the Argo Rollouts conf.yaml
file.

To see the conf.yaml file to get a better idea, see:
https://docs.datadoghq.com/integrations/argo-rollouts/\#configuration

So in order to actually have granular insight into canary and stable
pods, its should be applied on the conf.yaml file. The annotations here,
sadly, do not actualy contribute to argo_rollouts metrics provided by
the integration. It was worth a shot, however.

Signed-off-by: Eduardo Morales <emorales@nextdoor.com>